@@ -11,32 +11,34 @@
 	},
 	"type": "module",
 	"dependencies": {
-		"@vercel/analytics": "^1.3.1",
-		"zod": "^3.23.8"
+		"@tailwindcss/vite": "^4.0.0",
+		"@vercel/analytics": "^1.4.1",
+		"@vercel/speed-insights": "^1.1.0",
+		"zod": "^3.24.1"
 	},
 	"devDependencies": {
-		"@fontsource/fira-mono": "^5.0.14",
+		"@fontsource/fira-mono": "^5.1.1",
 		"@neoconfetti/svelte": "^2.2.1",
-		"@sveltejs/adapter-auto": "^3.2.4",
-		"@sveltejs/adapter-vercel": "^5.4.3",
-		"@sveltejs/kit": "^2.5.25",
-		"@sveltejs/vite-plugin-svelte": "^4.0.0-next.6",
-		"@tailwindcss/vite": "^4.0.0-alpha.20",
-		"@types/cookie": "^0.6.0",
-		"@typescript-eslint/eslint-plugin": "^8.3.0",
-		"@typescript-eslint/parser": "^8.3.0",
+		"@sveltejs/adapter-auto": "^4.0.0",
+		"@sveltejs/adapter-vercel": "^5.5.3",
+		"@sveltejs/kit": "^2.16.1",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@types/cookie": "^1.0.0",
+		"@typescript-eslint/eslint-plugin": "^8.21.0",
+		"@typescript-eslint/parser": "^8.21.0",
 		"autoprefixer": "^10.4.20",
-		"eslint": "^9.9.1",
-		"eslint-plugin-svelte": "^2.43.0",
-		"postcss": "^8.4.41",
-		"prettier": "^3.3.3",
-		"prettier-plugin-svelte": "^3.2.6",
-		"prettier-plugin-tailwindcss": "^0.6.6",
-		"svelte": "^4.2.19",
-		"svelte-check": "^3.8.6",
-		"tailwindcss": "^4.0.0-alpha.20",
-		"tslib": "^2.7.0",
-		"typescript": "^5.5.4",
-		"vite": "^5.4.2"
+		"eslint": "^9.18.0",
+		"eslint-plugin-svelte": "^2.46.1",
+		"prettier": "^3.4.2",
+		"prettier-plugin-svelte": "^3.3.3",
+		"prettier-plugin-tailwindcss": "^0.6.11",
+		"sharp": "^0.33.5",
+		"svelte": "^5.19.2",
+		"svelte-check": "^4.1.4",
+		"tailwindcss": "^4.0.0",
+		"tslib": "^2.8.1",
+		"typescript": "^5.7.3",
+		"vite": "^6.0.11",
+		"vite-plugin-image-optimizer": "^1.1.8"
 	}
 }
