@@ -2,9 +2,10 @@
   import { cartStoreData } from "$lib/cart";
   import DotLoading from "$components/DotLoading.svelte";
 
-  export let lineId: string;
-  let disabled: boolean;
-  async function updatCarItem() {
+  let { lineId }: { lineId: string } = $props();
+  let disabled = $state(false);
+
+  async function onclick() {
     disabled = true;
     const data = await fetch("/api/cart", {
       method: "DELETE",
@@ -21,8 +22,8 @@
 <button
   type="button"
   aria-label="Remover Item"
-  class="absolute z-20 -mt-2 ml-[68px] flex h-5 w-5 items-center justify-center rounded-full border border-neutral-900 bg-neutral-500 text-neutral-900"
-  on:click={updatCarItem}
+  class="absolute z-20 -mt-2 ml-[68px] flex h-5 w-5 cursor-pointer items-center justify-center rounded-full border border-neutral-900 bg-neutral-500 text-neutral-900"
+  {onclick}
   {disabled}
 >
   {#if disabled}
