@@ -1,8 +1,7 @@
 <script lang="ts">
-  import type { ActionData } from "./$types";
   import Input from "$components/Inputs/index.svelte";
 
-  export let form: ActionData;
+  let { form } = $props();
 </script>
 
 <svelte:head>
@@ -10,7 +9,7 @@
 </svelte:head>
 
 <section
-  class="m-auto my-12 max-w-md rounded-lg border border-neutral-800 bg-neutral-950 py-8 px-8"
+  class="m-auto my-12 max-w-md rounded-lg border border-neutral-800 bg-neutral-950 px-8 py-8"
 >
   <h1 class="mb-8 text-center text-3xl font-medium text-neutral-100">
     Ative a conta
@@ -37,7 +36,7 @@
       {#if form?.fields}<p class="text-red-400">{form?.message}</p>{/if}
     </span>
     <button
-      class="mt-6 ml-auto w-max rounded-full bg-blue-600 py-2 px-6 text-blue-50 hover:opacity-95"
+      class="mt-6 ml-auto w-max cursor-pointer rounded-full bg-blue-600 px-6 py-2 text-blue-50 hover:opacity-95"
       type="submit"
     >
       Ativar conta
