@@ -1,10 +1,9 @@
 <script lang="ts">
-  import type { ActionData } from "./$types";
   import DotLoading from "$components/DotLoading.svelte";
   import Input from "$components/Inputs/index.svelte";
 
-  export let form: ActionData;
-  let loading = false;
+  let { form } = $props();
+  let isLoading = $state(false);
 </script>
 
 <svelte:head>
@@ -12,7 +11,7 @@
 </svelte:head>
 
 <section
-  class="m-auto my-12 max-w-md rounded-lg border border-neutral-800 bg-neutral-950 py-8 px-8"
+  class="m-auto my-12 max-w-md rounded-lg border border-neutral-800 bg-neutral-950 px-8 py-8"
 >
   <h1 class="mb-8 text-center text-3xl font-medium text-neutral-100">
     Registrar
@@ -21,7 +20,7 @@
     method="POST"
     action="?/register"
     class="mb-4 flex flex-col"
-    on:submit={() => (loading = true)}
+    onsubmit={() => (isLoading = true)}
   >
     <Input
       id="name"
@@ -84,13 +83,13 @@
     </label>
     <button
       type="submit"
-      class="float-right mt-6 ml-auto w-max rounded-full bg-blue-600 text-blue-50 hover:opacity-95"
-      disabled={loading}
-      data-loading={loading}
+      class="float-right mt-6 ml-auto w-max cursor-pointer rounded-full bg-blue-600 text-blue-50 hover:opacity-95"
+      disabled={isLoading}
+      data-loading={isLoading}
       aria-label="Registre-se"
     >
-      <span class="block h-10 py-2 px-6">
-        {#if loading}
+      <span class="block h-10 px-6 py-2">
+        {#if isLoading}
           <DotLoading />
         {:else}
           Registre-se
