@@ -1,6 +1,6 @@
 <script lang="ts">
   import Card from "$components/Product/Card.svelte";
-  export let data;
+  let { data } = $props();
 </script>
 
 <svelte:head>
@@ -15,7 +15,7 @@
   <ul
     class="grid grid-flow-row grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-3"
   >
-    {#each data.products as product (product.handle)}
+    {#each data.products || [] as product (product.handle)}
       <li class="aspect-square w-full">
         <Card productProps={product} />
       </li>
