@@ -1,11 +1,11 @@
-<span class="m-auto flex h-full items-center justify-center gap-0.5">
+<span class="m-auto flex items-center justify-center gap-0.5">
   <span
-    class="animate-bouncedot size-1 rounded-full bg-neutral-50 [animation-delay:0.2s]"
-  />
+    class="animate-dotbounce size-1 rounded-full bg-neutral-50 [animation-delay:0.2s]"
+  ></span>
   <span
-    class="animate-bouncedot size-1 rounded-full bg-neutral-50 [animation-delay:0.3s]"
-  />
+    class="animate-dotbounce size-1 rounded-full bg-neutral-50 [animation-delay:0.3s]"
+  ></span>
   <span
-    class="animate-bouncedot size-1 rounded-full bg-neutral-50 [animation-delay:0.4s]"
-  />
+    class="animate-dotbounce size-1 rounded-full bg-neutral-50 [animation-delay:0.4s]"
+  ></span>
 </span>
