@@ -5,7 +5,7 @@
     format: string;
   }
 
-  function formatTextInput(event: Event & { currentTarget: HTMLInputElement }) {
+  function oninput(event: Event & { currentTarget: HTMLInputElement }) {
     const value = event.currentTarget?.value.replace(/\D/g, "");
     // /^(\d{2})(\d{1})(\d{4})(\d{0,4}).*/
     // "($1) $2 $3-$4"
@@ -25,8 +25,8 @@
   {/if}
   <input
     {...$$restProps}
-    class="h-10 w-full appearance-none rounded-lg border border-neutral-700 bg-transparent py-2 px-4 text-neutral-100 outline-none sm:text-sm [&:not(:placeholder-shown)]:invalid:border-red-500"
-    on:input={formatTextInput}
+    class="h-10 w-full appearance-none rounded-lg border border-neutral-700 bg-transparent px-4 py-2 text-neutral-100 outline-none sm:text-sm [&:not(:placeholder-shown)]:invalid:border-red-500"
+    {oninput}
   />
   {#if $$restProps.description}
     <p class="mt-1 block text-xs text-slate-500">{$$restProps.description}</p>
