@@ -1,9 +1,9 @@
 <script lang="ts">
-  import { page } from "$app/stores";
+  import { page } from "$app/state";
 </script>
 
 <svelte:head>
-  <title>{$page.error?.message} - {$page.status}</title>
+  <title>{page.error?.message} - {page.status}</title>
 </svelte:head>
 
 <h1
@@ -15,5 +15,5 @@
   class="mt-6 mx-auto flex w-fit items-center justify-between gap-2 bg-blue-600 px-4 text-3xl font-semibold text-neutral-100"
 >
   <span class="text-base">status:</span>
-  {$page.status}
+  {page.status}
 </p>
